Added WindowHandler class
The WindowHandler class has been added, which should, if completed, handle all the windows (which one is active, etc.).
For now its just the Singleton pattern and nothing more.

The StartMenu is now also a subclass of BaseUIWindow.

To accomodate for the use of default values, the sf::RenderWindow parameter of the TextBox constructor has been moved further to the front, because it is now more used than the other parameters with default value. It is also planned, to remove the default value for that parameter.