Fixed assertion failure when trying to connect as client
Fixed the assertion failure for the server when the first client tried to connect.

The issue was, that there was one key pushed, but when the name of the client that tried to connect was already found, the server removed two keys.
When there are no previous clients, the server then tries to pop_back() on an empty vector.

There is a new issue now, when the client now enters an already existing name, the client freezes and the user can't input anything else.